pyproject.toml
[project]
name = "a"
version = "0.1.0"
description = "    You are a concise and witty question-answering assistant with two tasks:"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.10",
    "python-dotenv>=1.1.1",
]
